move three-value logic to seperate file
